FIxed the same stuff the fith time :/

+++ Steckkarte_moc/src/adc.c
@@ -63,7 +63,7 @@ void ADC1_IRQHandler()
 {
 	uint32_t result=ADC1->DR;				//fetch conversion; clears the EOC flag
 	float voltage= (float)result*3.3/0xFFF0;
-	new_data(SPANNUNG_IN,(float)voltage);				//store conversion
+	new_data(SPANNUNG_IN,voltage);				//store conversion
 	new_data(SPANNUNG_OUT,1);
 	set_dac(1);
 }
